refactor(service): split global procmgr gate from per-process gates
Align GetServiceManagerType with main (cache + getServiceManagerType) and
apply the global procmgr gate before caching ProcmgrType. Move host-wide
gate helpers to service and DDOT process gate to procmgr (DDOTManaged).
Replace BaseServiceManagerType/UsesSystemd checks with GetServiceManagerType.
