+++ pkg/fleet/installer/packages/service/BUILD.bazel
@@ -2,14 +2,10 @@ load("@rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "service",
-    srcs = ["service.go"],
+    srcs = [
+        "gate.go",
+        "service.go",
+    ],
     importpath = "github.com/DataDog/datadog-agent/pkg/fleet/installer/packages/service",
     visibility = ["//visibility:public"],
-    # keep: platform select; procmgr dep is Linux-only, empty on Windows
-    deps = select({
-        "@rules_go//go/platform:windows": [],
-        "//conditions:default": [
-            "//pkg/fleet/installer/packages/service/procmgr",
-        ],
-    }),
 )
